Moving util.get_valid_https_verify to utilconfig
