+++ b2handle/tests/utilconfig_test.py
@@ -4,26 +4,26 @@ if sys.version_info < (2, 7):
 else:
     import unittest
 
-from b2handle import util
+from b2handle import utilconfig
 
 
-class UtilTestCase(unittest.TestCase):
+class UtilConfigTestCase(unittest.TestCase):
 
     def test_valid_https_verify_bool_true(self):
         """Test return bool True when getting bool True"""
-        self.assertEqual(util.get_valid_https_verify(True), True)
+        self.assertEqual(utilconfig.get_valid_https_verify(True), True)
 
     def test_valid_https_verify_string_true(self):
         """Test return bool True when getting string True"""
-        self.assertEqual(util.get_valid_https_verify('True'), True)
+        self.assertEqual(utilconfig.get_valid_https_verify('True'), True)
 
     def test_valid_https_verify_string_false(self):
         """Test return bool False when getting string False"""
-        self.assertEqual(util.get_valid_https_verify('False'), False)
+        self.assertEqual(utilconfig.get_valid_https_verify('False'), False)
 
     def test_valid_https_verify_bool_string(self):
         """Test return string when getting a string value in https_verify"""
         self.assertEqual(
-            util.get_valid_https_verify('ca_cert.crt'),
+            utilconfig.get_valid_https_verify('ca_cert.crt'),
             'ca_cert.crt'
         )
